fix(SPROUT-95): int, long 타입 불일치 수정

+++ src/main/java/com/example/sprout/domain/dashboard/dto/DashboardStatisticsResponse.java
@@ -5,7 +5,7 @@ public record DashboardStatisticsResponse(
         long receivedLikeCount,
         int attendanceStreak,
         int interviewCount,
-        long resumeCount
+        int resumeCount
 ) {
     public static DashboardStatisticsResponse of(
             int postCount, long receivedLikeCount, int attendanceStreak, int interviewCount, int resumeCount
